@@ -6,45 +6,42 @@
     "homepage": "https://larastreamers.com",
     "license": "MIT",
     "require": {
-        "php": "^8.2",
+        "php": "^8.3",
         "abraham/twitteroauth": "^5.0",
         "calebporzio/sushi": "^2.4",
         "doctrine/dbal": "^3.5",
         "guzzlehttp/guzzle": "^7.4",
         "laravel/framework": "^11.0",
-        "laravel/jetstream": "^5.0",
         "laravel/sanctum": "^4.0",
         "laravel/tinker": "^2.9",
         "league/flysystem-ftp": "^3.0",
-        "livewire/livewire": "^3.3",
+        "livewire/livewire": "^3.0",
         "owenvoke/blade-fontawesome": "^2.6",
         "spatie/data-transfer-object": "^3.7",
         "spatie/icalendar-generator": "^2.3",
         "spatie/laravel-backup": "^8.6",
         "spatie/laravel-feed": "^4.4",
         "spatie/laravel-ignition": "^2.4",
+        "spatie/laravel-login-link": "^1.2",
         "spatie/laravel-mailcoach-mailer": "^1.2",
         "symfony/http-client": "^7.0",
         "symfony/postmark-mailer": "^7.0",
         "vinkla/hashids": "^12.0"
     },
     "require-dev": {
-        "brianium/paratest": "^7.0",
         "christophrumpel/missing-livewire-assertions": "^2.5",
         "fakerphp/faker": "^1.23",
         "friendsofphp/php-cs-fixer": "^3.6",
         "laravel/sail": "^1.26",
         "mockery/mockery": "^1.6",
-        "nunomaduro/collision": "^6.1",
+        "nunomaduro/collision": "^8.0",
         "nunomaduro/larastan": "^2.8",
         "nunomaduro/patrol": "^1.1",
-        "pestphp/pest": "^2.23",
-        "pestphp/pest-plugin-laravel": "^2.2",
-        "pestphp/pest-plugin-parallel": "^1.0",
+        "pestphp/pest": "^2.0",
+        "pestphp/pest-plugin-laravel": "^2.0",
         "roave/security-advisories": "dev-latest",
         "spatie/laravel-ray": "^1.35",
-        "spatie/test-time": "^1.3",
-        "phpunit/phpunit": "^9.5.19"
+        "spatie/test-time": "^1.3"
     },
     "config": {
         "optimize-autoloader": true,
