@@ -8,22 +8,20 @@ use Illuminate\Foundation\Configuration\Middleware;
 return Application::configure(basePath: dirname(__DIR__))
     ->withProviders()
     ->withRouting(
-        web: __DIR__.'/../routes/web.php',
-        // api: __DIR__.'/../routes/api.php',
-        commands: __DIR__.'/../routes/console.php',
+        web: __DIR__ . '/../routes/web.php',
+        api: __DIR__ . '/../routes/api.php',
+        commands: __DIR__ . '/../routes/console.php',
         // channels: __DIR__.'/../routes/channels.php',
         health: '/up',
     )
     ->withMiddleware(function (Middleware $middleware) {
-        $middleware->redirectGuestsTo(fn () => route('login'));
+        $middleware->redirectGuestsTo(fn() => route('login'));
         $middleware->redirectUsersTo(AppServiceProvider::HOME);
-
-        $middleware->web(\Laravel\Jetstream\Http\Middleware\AuthenticateSession::class);
 
         $middleware->throttleApi();
     })
     ->withExceptions(function (Exceptions $exceptions) {
-        $exceptions->reportable(function(Throwable $e) {
+        $exceptions->reportable(function (Throwable $e) {
             //
         });
     })->create();
