@@ -2,6 +2,6 @@
     @include('pages.partials.header-home')
 
     <main class="{{ $upcomingStream ? 'md:-mt-16' : '' }} flex-1 text-white bg-gray-darkest">
-        <livewire:stream-list />
+        <livewire:stream-list/>
     </main>
 </x-main-layout>
