@@ -1,8 +1,8 @@
 <?php
 
-use App\Http\Livewire\SubmitYouTubeLiveStream;
+use App\Livewire\SubmitYouTubeLiveStream;
 
-it('includes the submission livewire component on all pages', function() {
+it('includes the submission livewire component on all pages', function () {
     $this->get(route('home'))
         ->assertSeeLivewire(SubmitYouTubeLiveStream::class);
 
