@@ -1,7 +1,7 @@
 <?php
 
-use App\Http\Livewire\ImportYouTubeChannel;
 use App\Jobs\ImportYoutubeChannelStreamsJob;
+use App\Livewire\ImportYouTubeChannel;
 use App\Models\Channel;
 use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Facades\Queue;
@@ -10,7 +10,7 @@ use Tests\Fakes\YouTubeResponses;
 
 uses(YouTubeResponses::class);
 
-it('adds channel to database', function() {
+it('adds channel to database', function () {
     // Arrange
     Queue::fake();
 
@@ -18,7 +18,7 @@ it('adds channel to database', function() {
     $this->assertDatabaseCount(Channel::class, 0);
 
     // Arrange & Act
-    Http::fake(fn () => Http::response($this->channelResponse()));
+    Http::fake(fn() => Http::response($this->channelResponse()));
 
     Livewire::test(ImportYouTubeChannel::class)
         ->set('youTubeChannelId', 'UCdtd5QYBx9MUVXHm7qgEpxA')
@@ -32,10 +32,10 @@ it('adds channel to database', function() {
     ]);
 });
 
-it('dispatches job to import upcoming streams', function() {
+it('dispatches job to import upcoming streams', function () {
     // Arrange
     Queue::fake();
-    Http::fake(fn () => Http::response($this->channelResponse()));
+    Http::fake(fn() => Http::response($this->channelResponse()));
 
     // Act
     Livewire::test(ImportYouTubeChannel::class)
@@ -46,10 +46,10 @@ it('dispatches job to import upcoming streams', function() {
     Queue::assertPushed(ImportYoutubeChannelStreamsJob::class);
 });
 
-it('shows success message', function() {
+it('shows success message', function () {
     // Arrange
     Queue::fake();
-    Http::fake(fn () => Http::response($this->channelResponse()));
+    Http::fake(fn() => Http::response($this->channelResponse()));
 
     // Act & Assert
     Livewire::test(ImportYouTubeChannel::class)
@@ -58,9 +58,9 @@ it('shows success message', function() {
         ->assertSee('Channel "UCdtd5QYBx9MUVXHm7qgEpxA" was added successfully.');
 });
 
-it('shows youtube client error message', function() {
+it('shows youtube client error message', function () {
     // Arrange
-    Http::fake(fn () => Http::response([], 500));
+    Http::fake(fn() => Http::response([], 500));
 
     // Arrange & Act & Assert
     Livewire::test(ImportYouTubeChannel::class)
@@ -69,10 +69,10 @@ it('shows youtube client error message', function() {
         ->assertSee('YouTube API error: 500');
 });
 
-it('clears form after successful import', function() {
+it('clears form after successful import', function () {
     // Arrange
     Queue::fake();
-    Http::fake(fn () => Http::response($this->channelResponse()));
+    Http::fake(fn() => Http::response($this->channelResponse()));
 
     // Act & Assert
     Livewire::test(ImportYouTubeChannel::class)
@@ -81,7 +81,7 @@ it('clears form after successful import', function() {
         ->assertSet('youTubeChannelId', '');
 });
 
-it('wires properties and methods', function() {
+it('wires properties and methods', function () {
     // Arrange & Act & Assert
     Livewire::test(ImportYouTubeChannel::class)
         ->assertPropertyWired('youTubeChannelId')
