@@ -5,7 +5,7 @@ use App\Models\Stream;
 use Carbon\Carbon;
 use Vinkla\Hashids\Facades\Hashids;
 
-it('shows only finished streams', function() {
+it('shows only finished streams', function () {
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -32,7 +32,7 @@ it('shows only finished streams', function() {
         ->assertSee('Finished stream');
 });
 
-it('uses actual start time first for order if given', function() {
+it('uses actual start time first for order if given', function () {
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -58,7 +58,7 @@ it('uses actual start time first for order if given', function() {
         ]);
 });
 
-it('orders streams from latest to oldest', function() {
+it('orders streams from latest to oldest', function () {
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -90,7 +90,7 @@ it('orders streams from latest to oldest', function() {
         ]);
 });
 
-it('does not show deleted streams', function() {
+it('does not show deleted streams', function () {
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -111,22 +111,22 @@ it('does not show deleted streams', function() {
         ->assertDontSee('Stream deleted');
 });
 
-it('shows duration of stream if given', function() {
+it('shows duration of stream if given', function () {
     // Arrange
-    Stream::factory()
+    $stream = Stream::factory()
         ->withChannel()
         ->finished()
         ->withActualStartTime(Carbon::yesterday())
         ->withActualEndTime(Carbon::yesterday()->addHour()->addMinutes(12))
         ->create();
-
+    
     // Act & Assert
     $this
         ->get(route('archive'))
         ->assertSee('1h 12m');
 });
 
-it('searches for streams on title', function() {
+it('searches for streams on title', function () {
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -156,7 +156,7 @@ it('searches for streams on title', function() {
         ]);
 });
 
-it('searches for streams on channel title', function() {
+it('searches for streams on channel title', function () {
     // Arrange
     $channelShown = Channel::factory()->create(['name' => 'Channel Shown']);
     Stream::factory()
@@ -192,7 +192,7 @@ it('searches for streams on channel title', function() {
 it('searches
  for streams by specific
  ->$this->withTitle()
-  streamer', function() {
+  streamer', function () {
     // Arrange
     Stream::factory()
         ->withChannel()
@@ -212,7 +212,7 @@ it('searches
         ->assertDontSee('Stream Hidden');
 });
 
-it('searches for streams by specific streamer and search term', function() {
+it('searches for streams by specific streamer and search term', function () {
     // Arrange
     $channel = Channel::factory()->create();
     Stream::factory()
@@ -233,7 +233,7 @@ it('searches for streams by specific streamer and search term', function() {
         ->assertDontSee('Stream Hidden');
 });
 
-it('orders streamers by name', function() {
+it('orders streamers by name', function () {
     $this->withoutExceptionHandling();
     // Arrange
     Channel::factory()
