@@ -1,26 +1,26 @@
 <?php
 
-it('can show the home page', function() {
+it('can show the home page', function () {
     $this->get(route('home'))
-         ->assertOk();
+        ->assertOk();
 });
 
-it('can show the feed', function() {
+it('can show the feed', function () {
     $this->get('/feed')
         ->assertOk();
 });
 
-it('can show the archive', function() {
+it('can show the archive', function () {
     $this->get(route('archive'))
         ->assertOk();
 });
 
-it('can show the calendar page', function() {
+it('can show the calendar page', function () {
     $this->get(route('calendar.ics'))
         ->assertOk();
 });
 
-it('can privacy policy page', function() {
+it('can privacy policy page', function () {
     $this->get(route('privacy.policy'))
         ->assertOk();
 });
